+++ tests/integration/it_utils.py
@@ -321,6 +321,7 @@ def _get_non_decorator_code(function):
 
 def create_amm_pool(
     client: SyncClient = JSON_RPC_CLIENT,
+    enable_amm_clawback: bool = False,
 ) -> Dict[str, Any]:
     issuer_wallet = Wallet.create()
     fund_wallet(issuer_wallet)
@@ -337,6 +338,16 @@ def create_amm_pool(
         issuer_wallet,
     )
 
+    # The below flag is required for AMMClawback tests
+    if enable_amm_clawback:
+        sign_and_reliable_submission(
+            AccountSet(
+                account=issuer_wallet.classic_address,
+                set_flag=AccountSetAsfFlag.ASF_ALLOW_TRUSTLINE_CLAWBACK,
+            ),
+            issuer_wallet,
+        )
+
     sign_and_reliable_submission(
         TrustSet(
             account=lp_wallet.classic_address,
@@ -388,11 +399,13 @@ def create_amm_pool(
         "asset": asset,
         "asset2": asset2,
         "issuer_wallet": issuer_wallet,
+        "lp_wallet": lp_wallet,
     }
 
 
 async def create_amm_pool_async(
     client: AsyncClient = ASYNC_JSON_RPC_CLIENT,
+    enable_amm_clawback: bool = False,
 ) -> Dict[str, Any]:
     issuer_wallet = Wallet.create()
     await fund_wallet_async(issuer_wallet)
@@ -409,6 +422,16 @@ async def create_amm_pool_async(
         issuer_wallet,
     )
 
+    # The below flag is required for AMMClawback tests
+    if enable_amm_clawback:
+        await sign_and_reliable_submission_async(
+            AccountSet(
+                account=issuer_wallet.classic_address,
+                set_flag=AccountSetAsfFlag.ASF_ALLOW_TRUSTLINE_CLAWBACK,
+            ),
+            issuer_wallet,
+        )
+
     await sign_and_reliable_submission_async(
         TrustSet(
             account=lp_wallet.classic_address,
@@ -460,6 +483,7 @@ async def create_amm_pool_async(
         "asset": asset,
         "asset2": asset2,
         "issuer_wallet": issuer_wallet,
+        "lp_wallet": lp_wallet,
     }