+++ xrpl/models/transactions/__init__.py
@@ -11,6 +11,7 @@ from xrpl.models.transactions.account_set import (
     AccountSetFlagInterface,
 )
 from xrpl.models.transactions.amm_bid import AMMBid, AuthAccount
+from xrpl.models.transactions.amm_clawback import AMMClawback
 from xrpl.models.transactions.amm_create import AMMCreate
 from xrpl.models.transactions.amm_delete import AMMDelete
 from xrpl.models.transactions.amm_deposit import (
@@ -119,6 +120,7 @@ __all__ = [
     "AccountSetFlag",
     "AccountSetFlagInterface",
     "AMMBid",
+    "AMMClawback",
     "AMMCreate",
     "AMMDelete",
     "AMMDeposit",